requirements.txt
discord.py~=2.4.0
audioop-lts==0.2.1; python_version>='3.13'
eyed3~=0.9.7
pillow @ git+https://github.com/python-pillow/Pillow.git@main
numpy~=1.26.2
SQLAlchemy~=1.4.40
python-dotenv~=1.0.0
PyNaCl~=1.5.0
requests~=2.31.0
emoji~=2.6.0
alembic[tz]~=1.11.1
dateparser==1.2.0
aiohttp~=3.11.11
aiodns~=3.4.0